chore: refactor structure of Home module
