+++ suite-native/module-home/src/screens/HomeScreen/components/BiometricsBottomSheet.tsx
