+++ suite-native/module-home/src/screens/HomeScreen/components/EmptyConnectedDeviceState.tsx
