+++ suite-native/module-home/src/screens/HomeScreen/components/PortfolioContent.tsx
