+++ suite-native/module-home/src/screens/HomeScreen/components/PortfolioGraph.tsx
