+++ suite-native/module-home/src/screens/HomeScreen/components/PortfolioGraphHeader.tsx
